refactor/move fuzz playground server to pkg
